refactor(provider): move passport projection startup runner to infrastructure
